fix tests & examples

+++ examples/hist_test/hist_test.ino
@@ -13,7 +13,7 @@
 float b[] = {
   0, 100, 200, 300, 325, 350, 375 };
 
-Histogram<int16_t> hist(7, b);
+Histogram<uint32_t> hist(7, b);
 
 uint32_t lastTime = 0;
 const uint32_t threshold = 25;  // milliseconds, for updating display
